added generic function for daterange and removed pandas dependency from session.py
